requirements.txt
RPi.GPIO==0.6.3

